Added SparkConnection Trait, moved loop into syncLoop function, added Scala exception handling
